Refactored. Converted to Maven, Organized, Updated.
